15_Intermediate To JS
-15.0 Random Number Generation in Javascript: Building a Love Calculator:
*Math.random(); => Random number between 0,1(not including 1) with 16 bits size.

-15.1 Control Statements:(IF-Else):
BMI <18.5, the output should be: "Your BMI is <bmi>, so you are underweight."

BMI 18.5-24.9, the output should be "Your BMI is <bmi>, so you have a normal weight."

BMI >24.9, the output should be "Your BMI is <bmi>, so you are overweight."

-15.2 BMI Calculator Advanced (IF/ELSE):

-15.3 Leap Year Challenge Exercise:
This is how to work out whether if a particular year is a leap year:
A year is a leap year if it is evenly divisible by 4 ;
except if that year is also evenly divisible by 100;
unless that year is also evenly divisible by 400.

-15.4 Check List Challenge :

-15.5 Print 1 To 100 m3 => fizz, m5 => buzz.

-15.6 Who's Buying Lunch? Code Challenge:
You are going to write a function which will select a random name from a list of names. The person selected will have to pay for everybody's food bill.
Example Input:
["Angela", "Ben", "Jenny", "Michael", "Chloe"]

Example Output:
Michael is going to buy lunch today!

-15.7 99 Bottles Song:
99 bottles of beer on the wall, 99 bottles of beer.
Take one down and pass it around, 98 bottles of beer on the wall.

-15.8 Fibonacci Generator Challenge:
fib(5) 0 1 1 2 3